refactor: restructure plugin to utilize more standardized class-based organization with built-in loader
