+++ includes/class-loader.php
@@ -1,0 +1,122 @@
+<?php
+/**
+ * Registers the Loader class.
+ *
+ * @package SLO\Loader
+ * @since 0.0.1
+ */
+
+namespace SLO;
+
+/**
+ * Register all actions and filters for the plugin.
+ *
+ * Maintain a list of all hooks that are registered throughout
+ * the plugin, and register them with the WordPress API. Call the
+ * run function to execute the list of actions and filters.
+ *
+ * @package SLO
+ * @subpackage SLO/Loader
+ */
+class Loader {
+
+  /**
+   * The array of actions registered with WordPress.
+   *
+   * @var array $actions    The actions registered with WordPress to fire when the plugin loads.
+   *
+   * @access protected
+   * @since 0.0.1
+   */
+  protected $actions;
+
+  /**
+   * The array of filters registered with WordPress.
+   *
+   * @var array $filters    The filters registered with WordPress to fire when the plugin loads.
+   *
+   * @access protected
+   * @since 0.0.1
+   */
+  protected $filters;
+
+  /**
+   * Initialize the collections used to maintain the actions and filters.
+   *
+   * @since 0.0.1
+   */
+  public function __construct() {
+    $this->actions = array();
+    $this->filters = array();
+  }
+
+  /**
+   * Add a new action to the collection to be registered with WordPress.
+   *
+   * @param string $hook             The name of the WordPress action that is being registered.
+   * @param object $component        A reference to the instance of the object on which the action is defined.
+   * @param string $callback         The name of the function definition on the $component.
+   * @param int    $priority         Optional. he priority at which the function should be fired. Default is 10.
+   * @param int    $accepted_args    Optional. The number of arguments that should be passed to the $callback. Default is 1.
+   *
+   * @since 0.0.1
+   */
+  public function add_action( $hook, $component, $callback, $priority = 10, $accepted_args = 1 ) {
+    $this->actions = $this->add( $this->actions, $hook, $component, $callback, $priority, $accepted_args );
+  }
+
+  /**
+   * Add a new filter to the collection to be registered with WordPress.
+   *
+   * @param string $hook             The name of the WordPress filter that is being registered.
+   * @param object $component        A reference to the instance of the object on which the filter is defined.
+   * @param string $callback         The name of the function definition on the $component.
+   * @param int    $priority         Optional. he priority at which the function should be fired. Default is 10.
+   * @param int    $accepted_args    Optional. The number of arguments that should be passed to the $callback. Default is 1.
+   *
+   * @since 0.0.1
+   */
+  public function add_filter( $hook, $component, $callback, $priority = 10, $accepted_args = 1 ) {
+    $this->filters = $this->add( $this->filters, $hook, $component, $callback, $priority, $accepted_args );
+  }
+
+  /**
+   * A utility function that is used to register the actions and hooks into a single
+   * collection.
+   *
+   * @param array  $hooks            The collection of hooks that is being registered (that is, actions or filters).
+   * @param string $hook             The name of the WordPress filter that is being registered.
+   * @param object $component        A reference to the instance of the object on which the filter is defined.
+   * @param string $callback         The name of the function definition on the $component.
+   * @param int    $priority         The priority at which the function should be fired.
+   * @param int    $accepted_args    The number of arguments that should be passed to the $callback.
+   * @return array                   The collection of actions and filters registered with WordPress.
+   *
+   * @access private
+   * @since 0.0.1
+   */
+  private function add( $hooks, $hook, $component, $callback, $priority, $accepted_args ) {
+    $hooks[] = array(
+      'hook'          => $hook,
+      'component'     => $component,
+      'callback'      => $callback,
+      'priority'      => $priority,
+      'accepted_args' => $accepted_args,
+    );
+    return $hooks;
+  }
+
+  /**
+   * Register the filters and actions with WordPress.
+   *
+   * @since 0.0.1
+   */
+  public function run() {
+    foreach ( $this->filters as $hook ) {
+      add_filter( $hook['hook'], array( $hook['component'], $hook['callback'] ), $hook['priority'], $hook['accepted_args'] );
+    }
+    foreach ( $this->actions as $hook ) {
+      add_action( $hook['hook'], array( $hook['component'], $hook['callback'] ), $hook['priority'], $hook['accepted_args'] );
+    }
+  }
+}
