+++ templates/template-social-link-optimizer.php
@@ -30,27 +30,27 @@ get_header();
         <div class="entry-content gpa-social-link-optimizer">
     
           <?php
-          $is_gpaslo_page_template = is_page_template( '/templates/template-social-link-optimizer.php' );
+          $is_gpaslo_page_template = is_page_template( 'template-social-link-optimizer.php' );
 
           if ( is_search() || ! is_singular() && 'summary' === get_theme_mod( 'blog_content', 'full' ) ) {
             the_excerpt();
           } elseif ( $is_gpaslo_page_template ) {
             $all_settings = get_option( 'gpalab_slo_settings_option_name' );
-            $display_as = $all_settings['display_gpalab_slo_as_a_0'];
-            $layout = ( isset( $display_as ) && $display_as !== '' )
+            $display_as   = $all_settings['display_gpalab_slo_as_a_0'];
+            $layout       = ( isset( $display_as ) && '' !== $display_as )
               ? $display_as
               : 'grid';
 
             $social_accts = array(
-              'twitter'  => $all_settings['twitter_feed_3'],
-              'facebook' => $all_settings['facebook_page_1'],
+              'twitter'   => $all_settings['twitter_feed_3'],
+              'facebook'  => $all_settings['facebook_page_1'],
               'instagram' => $all_settings['instagram_feed_5'],
-              'youtube'  => $all_settings['youtube_channel_4'],
-              'linkedin' => $all_settings['linkedin_profile_2'],
+              'youtube'   => $all_settings['youtube_channel_4'],
+              'linkedin'  => $all_settings['linkedin_profile_2'],
             );
 
             $assets_dir = plugins_url( 'social-link-optimizer' ) . '/assets/';
-          ?>
+            ?>
 
             <aside>
               <h3 id="social-accts" class="hide-visually">
@@ -58,16 +58,16 @@ get_header();
               </h3>
               <ul class="social-media-accts" aria-describedby="social-accts">
               <?php
-                foreach ( $social_accts as $key => $value ) {
-                  if ( isset( $value ) && $value !== '' ) {
-                    echo '<li>';
-                    echo '<a href="' . $value . '">';
-                    echo '<img src="' . $assets_dir . $key . '.svg" alt="" height="24" width="24">';
-                    echo '<span class="hide-visually">' . $key . '</span>';
-                    echo '</a>';
-                    echo '</li>';
-                  }
+              foreach ( $social_accts as $key => $value ) {
+                if ( isset( $value ) && '' !== $value ) {
+                  echo '<li>';
+                  echo '<a href="' . $value . '">';
+                  echo '<img src="' . $assets_dir . $key . '.svg" alt="" height="24" width="24">';
+                  echo '<span class="hide-visually">' . $key . '</span>';
+                  echo '</a>';
+                  echo '</li>';
                 }
+              }
               ?>
               </ul>
             </aside>
@@ -84,9 +84,9 @@ get_header();
               if ( $the_query->have_posts() ) {
                 while ( $the_query->have_posts() ) {
                   $the_query->the_post();
-                  $anchor_tag_open = '<a href="' . esc_url( get_permalink() ) . '">';
+                  $anchor_tag_open  = '<a href="' . esc_url( get_permalink() ) . '">';
                   $anchor_tag_close = '</a>';
-              
+
                   echo '<li>';
                   echo '<article>';
                   echo '<h3 class="gpa-social-title">';
@@ -95,7 +95,7 @@ get_header();
                   echo $layout == 'list' ? $anchor_tag_close : null;
                   echo '</h3>';
                   echo $layout == 'grid' ? $anchor_tag_open : null;
-                  $layout == 'grid' ? the_post_thumbnail( 'post-thumbnail', ['class' => 'gpa-social-thumbnail'] ) : null;
+                  $layout == 'grid' ? the_post_thumbnail( 'post-thumbnail', array( 'class' => 'gpa-social-thumbnail' ) ) : null;
                   echo $layout == 'grid' ? $anchor_tag_close : null;
                   echo '</article>';
                   echo '</li>';
